+++ actix-http/src/header/into_pair.rs
@@ -1,19 +1,17 @@
-//! [`IntoHeaderPair`] trait and implementations.
+//! [`TryIntoHeaderPair`] trait and implementations.
 
 use std::convert::TryFrom as _;
 
-use http::{
-    header::{HeaderName, InvalidHeaderName, InvalidHeaderValue},
-    Error as HttpError, HeaderValue,
+use super::{
+    Header, HeaderName, HeaderValue, IntoHeaderValue, InvalidHeaderName, InvalidHeaderValue,
 };
+use crate::error::HttpError;
 
-use super::{Header, IntoHeaderValue};
-
-/// An interface for types that can be converted into a [`HeaderName`]/[`HeaderValue`] pair for
+/// An interface for types that can be converted into a [`HeaderName`] + [`HeaderValue`] pair for
 /// insertion into a [`HeaderMap`].
 ///
 /// [`HeaderMap`]: super::HeaderMap
-pub trait IntoHeaderPair: Sized {
+pub trait TryIntoHeaderPair: Sized {
     type Error: Into<HttpError>;
 
     fn try_into_header_pair(self) -> Result<(HeaderName, HeaderValue), Self::Error>;
@@ -34,7 +32,7 @@ impl From<InvalidHeaderPart> for HttpError {
     }
 }
 
-impl<V> IntoHeaderPair for (HeaderName, V)
+impl<V> TryIntoHeaderPair for (HeaderName, V)
 where
     V: IntoHeaderValue,
     V::Error: Into<InvalidHeaderValue>,
@@ -50,7 +48,7 @@ where
     }
 }
 
-impl<V> IntoHeaderPair for (&HeaderName, V)
+impl<V> TryIntoHeaderPair for (&HeaderName, V)
 where
     V: IntoHeaderValue,
     V::Error: Into<InvalidHeaderValue>,
@@ -66,7 +64,7 @@ where
     }
 }
 
-impl<V> IntoHeaderPair for (&[u8], V)
+impl<V> TryIntoHeaderPair for (&[u8], V)
 where
     V: IntoHeaderValue,
     V::Error: Into<InvalidHeaderValue>,
@@ -83,7 +81,7 @@ where
     }
 }
 
-impl<V> IntoHeaderPair for (&str, V)
+impl<V> TryIntoHeaderPair for (&str, V)
 where
     V: IntoHeaderValue,
     V::Error: Into<InvalidHeaderValue>,
@@ -100,22 +98,24 @@ where
     }
 }
 
-impl<V> IntoHeaderPair for (String, V)
+impl<V> TryIntoHeaderPair for (String, V)
 where
     V: IntoHeaderValue,
     V::Error: Into<InvalidHeaderValue>,
 {
     type Error = InvalidHeaderPart;
 
+    #[inline]
     fn try_into_header_pair(self) -> Result<(HeaderName, HeaderValue), Self::Error> {
         let (name, value) = self;
         (name.as_str(), value).try_into_header_pair()
     }
 }
 
-impl<T: Header> IntoHeaderPair for T {
+impl<T: Header> TryIntoHeaderPair for T {
     type Error = <T as IntoHeaderValue>::Error;
 
+    #[inline]
     fn try_into_header_pair(self) -> Result<(HeaderName, HeaderValue), Self::Error> {
         Ok((T::name(), self.try_into_value()?))
     }