+++ src/response/mod.rs
@@ -1,9 +1,13 @@
 mod builder;
+mod customize_responder;
 mod http_codes;
+mod responder;
 #[allow(clippy::module_inception)]
 mod response;
 
 pub use self::builder::HttpResponseBuilder;
+pub use self::customize_responder::CustomizeResponder;
+pub use self::responder::Responder;
 pub use self::response::HttpResponse;
 
 #[cfg(feature = "cookies")]